fixed bugs with imports, renamed review component
